refactor: types/ folder dissolved — MockUserIds and chat-owned SignalrResult
"Types" was a junk-drawer folder+namespace with no counterpart anywhere else in
the repo — everything is a type. Both occupants re-homed by MEANING:

- UserIds → MockUserIds (web-contracts/mocks/): both consumers are mock-mode
  (SPA mock auth, contract mock factories) and nothing seeds users — the Purpose
  region's "seed data" claim was false and is corrected. Namespace is root
  TimeWarp.Architecture, matching the AuthorizationConstants well-known-constants
  precedent; fields now readonly.
- SignalrResult → web-contracts/features/chat/: generic by shape but chat-owned
  by residence (the chat hub is its only consumer). Design region records the
  promotion trigger: a second hub borrowing it will trip TWPA0009 and force the
  shared-home decision explicitly.

TimeWarp.Architecture.Types is gone; all seven stale usings purged (one left an
empty #if(web) guard in timewarp-testing — removed).

Maintainer ruling recorded: namespaces exist to avoid collisions and carry
meaning — folder organization must never drive namespace renames, and grab-bag
namespaces (Types/Common/Shared) are worse than specific ones.

Verified: dev build 0/0; default generation ships both files at their new homes.
