+++ source/container-apps/web/web-contracts/global-usings.cs
@@ -16,6 +16,5 @@ global using TimeWarp.Foundation;
 global using TimeWarp.Architecture.Features;
 global using TimeWarp.Foundation.Features;
 // Solution usings
-global using TimeWarp.Architecture.Types;
 global using TimeWarp.Foundation.Types;
 global using TimeWarp.Mediator;
