+++ tests/container-apps/web/web-server-integration-tests/GlobalUsings.cs
@@ -11,6 +11,5 @@ global using TimeWarp.Architecture.Configuration;
 global using TimeWarp.Foundation.Configuration;
 global using TimeWarp.Architecture.Features.Analytics;
 global using TimeWarp.Architecture.Testing;
-global using TimeWarp.Architecture.Types;
 global using TimeWarp.Foundation.Types;
 global using TimeWarp.Fixie;
